requirements.txt
annotated-types==0.7.0
anyio==4.10.0
arxiv==2.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.116.1
feedparser==6.0.12
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
PyMuPDF==1.26.4
requests==2.32.5
sgmllib3k==1.0.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0

# === 추가된 필수 패키지 ===
anthropic>=0.40.0
python-dotenv>=1.0.1
redis>=5.0.0
rq>=1.15.0
graphviz>=0.20.3
matplotlib>=3.9.0
pyyaml>=6.0.2

# === 선택 패키지 (추가 기능용) ===
pillow>=10.0.0
reportlab>=4.2.0
prometheus-client>=0.20.0
